+++ pages/[[server]]/list/[list]/accounts.vue
@@ -1,7 +1,10 @@
 <script setup lang="ts">
-const { listId } = defineProps<{
-  listId: string
-}>()
+definePageMeta({
+  name: 'list-accounts',
+})
+
+const params = useRoute().params
+const listId = $(computedEager(() => params.list as string))
 
 const paginator = useMastoClient().v1.lists.listAccounts(listId)
 </script>
